[chore] move schema to types
